requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
numpy==1.21.6
scikit-learn==1.0.2
Pillow==9.0.1
opencv-python-headless==4.5.5.64
joblib==1.1.0
gunicorn==20.1.0
tensorflow-cpu==2.11.0